clean memory type table implementation and plug into coarse grain memory tracking machinery
